--- src/camera/mod.rs
@@ -1,1 +1,0 @@
-pub mod menu_camera;
