+++ src/lib.rs
@@ -1,3 +1,2 @@
-pub mod camera;
 pub mod game;
 pub mod menu;
