+++ src/main.rs
@@ -1,10 +1,6 @@
 use bevy::prelude::*;
 use bevy::window::WindowResolution;
 use bevy::winit::WinitSettings;
-use lunarpunk::{
-    camera::menu_camera::spawn_camera, menu::button_interactions::main_menu_button_interaction,
-    menu::menu_scene::setup,
-};
 
 fn main() {
     App::new()
@@ -17,7 +13,6 @@ fn main() {
             ..default()
         }))
         .insert_resource(WinitSettings::desktop_app())
-        .add_systems(Startup, (setup, spawn_camera))
-        .add_systems(Update, main_menu_button_interaction)
+        .add_plugins(lunarpunk::menu::MenuPlugin)
         .run();
 }
