+++ src/menu/main/components.rs
@@ -1,0 +1,9 @@
+use bevy::prelude::Component;
+
+#[derive(Component, Debug)]
+pub enum MenuButtonAction {
+    Play,
+    Settings,
+    Credits,
+    Quit,
+}
