+++ src/menu/mod.rs
@@ -1,2 +1,15 @@
-pub mod button_interactions;
-pub mod menu_scene;
+use crate::menu::camera::CameraPlugin;
+use crate::menu::main::LayoutPlugin;
+use bevy::prelude::*;
+
+pub mod camera;
+mod components;
+pub mod main;
+
+pub struct MenuPlugin;
+
+impl Plugin for MenuPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_plugins((LayoutPlugin, CameraPlugin));
+    }
+}
